+++ lib/flame_camera_tools.dart
@@ -1,5 +1,5 @@
 library;
 
-export 'src/behaviors/smooth_follow_behavior.dart';
+export 'src/behaviors/behaviors.dart';
 export 'src/effects/shake_effect.dart';
 export 'src/extensions/camera_extensions.dart';