+++ lib/src/behaviors/behaviors.dart
@@ -1,0 +1,2 @@
+export 'deadzone.dart';
+export 'smooth_follow_behavior.dart';