+++ lib/src/extensions/camera_extensions.dart
@@ -6,26 +6,29 @@ import 'package:flame_camera_tools/flame_camera_tools.dart';
 import 'package:flutter/widgets.dart';
 
 extension FlameCameraTools on CameraComponent {
-  /// Smoothly follows the [target] with adjustable [stiffness],
-  /// allowing for control over the following behavior.
+  /// Smoothly follows the [target] with adjustable [stiffness], allowing fine
+  /// control over the following behavior.
   ///
-  /// The camera will gradually follow the [target] based on the [stiffness] parameter,
-  /// with the option to restrict following to only the horizontal or vertical axis.
-  /// The [deadZone] defines an area within which the camera will not follow the [target].
-  /// The [snap] parameter allows the camera to instantly align with the target's position
-  /// before starting the smooth following behavior.
+  /// The camera gradually follows the [target] based on [stiffness], optionally
+  /// restricting movement to horizontal or vertical axes. The [deadZone] defines
+  /// an area around the camera where it won't follow the target.
+  ///
+  /// If [snap] is true, the camera immediately aligns with the target before
+  /// starting smooth following.
   ///
   /// Parameters:
   /// - [target]: The position provider to follow.
-  /// - [stiffness]: The responsiveness of the camera's movement. The value should be between 0.0 and 1.0, where 0.0 means no movement and 1.0 means immediate following.
-  /// - [deadZone]: The area within which the camera will not follow the [target]. Defaults to `Rect.zero`.
-  /// - [horizontalOnly]: If true, restricts the following behavior to the horizontal axis.
-  /// - [verticalOnly]: If true, restricts the following behavior to the vertical axis.
-  /// - [snap]: If true, the camera will immediately align with the target's position before following.
+  /// - [stiffness]: Responsiveness of the follow behavior, from 0.0 (no movement)
+  ///   to 1.0 (instant snap).
+  /// - [deadZone]: Area inside which the camera doesn't follow the target.
+  ///   If not provided, a [CircularDeadzone] with radius `0` is used.
+  /// - [horizontalOnly]: If true, follow only along the x-axis.
+  /// - [verticalOnly]: If true, follow only along the y-axis.
+  /// - [snap]: If true, immediately jump to the target's position.
   void smoothFollow(
     ReadOnlyPositionProvider target, {
     double stiffness = 1.0,
-    Rect deadZone = Rect.zero,
+    Deadzone? deadZone,
     bool horizontalOnly = false,
     bool verticalOnly = false,
     bool snap = false,